fix(ui): code refactoring
