--- src/table.css
@@ -1,9 +1,0 @@
-table {
-    border-spacing: 0;
-    border: 1px solid black;
-
-}
-
-    
-
-
